Change the way rewarded video ads are loaded
Old way: Only try to load an ad and play when loaded when user clicks the
button, in case of error display a 'not available' toast but keep the
button enabled. Clicking the button again repeats the process.

New way: try to load an ad and display when loaded when user clicks the
button for the very first time, keep preloading ads after first video is
played. In case of load error show the toast but disable the button so
no more ads could be loaded.